refactor(Personagem): CRUD/Testes
Refafora service de Personagem para respeitar SOLID e adapta controller e testes para a refatoracao
