+++ src/main/java/br/com/batalharepg/avanade/service/personagem/CreatePersonagemService.java
@@ -1,8 +1,7 @@
-package br.com.batalharepg.avanade.service;
+package br.com.batalharepg.avanade.service.personagem;
 
 import br.com.batalharepg.avanade.configuration.PersonagemFactoryConfiguration;
 import br.com.batalharepg.avanade.dto.request.PersonagemRequest;
-import br.com.batalharepg.avanade.dto.request.PersonagemUpdateRequest;
 import br.com.batalharepg.avanade.dto.response.PersonagemResponse;
 import br.com.batalharepg.avanade.entities.Personagem;
 import br.com.batalharepg.avanade.exceptions.NomeExistenteException;
@@ -12,12 +11,9 @@ import br.com.batalharepg.avanade.repository.PersonagemRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.UUID;
-
 @Service
 @RequiredArgsConstructor
-public class PersonagemService {
+public class CreatePersonagemService {
     private final PersonagemRepository personagemRepository;
     private final PersonagemFactoryConfiguration factoryConfiguration;
 
@@ -31,30 +27,5 @@ public class PersonagemService {
             return personagemRepository.save(personagem).getResponseDto();
         }
         throw new NotFoundException("Tipo de personagem não encontrado");
-    }
-
-    public List<PersonagemResponse> listaTodosPersonagens() {
-        return personagemRepository.findAll().stream().map(Personagem::getResponseDto).toList();
-    }
-
-    public PersonagemResponse buscaPersonagemPorNome(String nome) {
-        return personagemRepository.findByNomeOrThrowException(nome)
-            .getResponseDto();
-    }
-
-    public PersonagemResponse buscaPersonagemPorUuid(UUID uuid) throws NotFoundException {
-        return personagemRepository.findByIdWithExceptionIfNotFound(uuid)
-            .getResponseDto();
-    }
-
-    public PersonagemResponse atualizaPersonagem(UUID uuid, PersonagemUpdateRequest dto) throws NotFoundException {
-        Personagem personagem = personagemRepository.findByIdWithExceptionIfNotFound(uuid);
-        personagem.setNome(dto.nome());
-        return personagemRepository.save(personagem).getResponseDto();
-    }
-
-    public void deletaPersonagem(UUID uuid) throws NotFoundException {
-        Personagem personagem = personagemRepository.findByIdWithExceptionIfNotFound(uuid);
-        personagemRepository.delete(personagem);
     }
 }
