--- src/test/java/br/com/batalharepg/avanade/service/PersonagemServiceTest.java
@@ -1,178 +1,0 @@
-package br.com.batalharepg.avanade.service;
-
-import br.com.batalharepg.avanade.configuration.PersonagemFactoryConfiguration;
-import br.com.batalharepg.avanade.dto.request.PersonagemRequest;
-import br.com.batalharepg.avanade.dto.request.PersonagemUpdateRequest;
-import br.com.batalharepg.avanade.dto.response.PersonagemResponse;
-import br.com.batalharepg.avanade.entities.Personagem;
-import br.com.batalharepg.avanade.exceptions.NomeExistenteException;
-import br.com.batalharepg.avanade.exceptions.NotFoundException;
-import br.com.batalharepg.avanade.factory.personagem.PersonagemFactory;
-import br.com.batalharepg.avanade.factory.personagem.TipoClassePersonagem;
-import br.com.batalharepg.avanade.factory.personagem.implementacoes.BarbaroFactory;
-import br.com.batalharepg.avanade.factory.personagem.implementacoes.OrcFactory;
-import br.com.batalharepg.avanade.repository.PersonagemRepository;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentMatchers;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.stubbing.Answer;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@SpringBootTest
-class PersonagemServiceTest {
-    @Mock
-    private PersonagemRepository personagemRepository;
-    @Mock
-    private PersonagemFactoryConfiguration factoryConfiguration;
-    @InjectMocks
-    private PersonagemService personagemService;
-
-    @BeforeEach
-    public void setUp() {
-        Mockito.when(factoryConfiguration.getFactory(ArgumentMatchers.any(TipoClassePersonagem.class)))
-            .thenAnswer((Answer<PersonagemFactory>) invocation -> {
-                TipoClassePersonagem argument = invocation.getArgument(0);
-
-                if (argument == TipoClassePersonagem.BARBARO) {
-                    return new BarbaroFactory();
-                } else if (argument == TipoClassePersonagem.ORC) {
-                    return new OrcFactory();
-                }
-                return null;
-            });
-    }
-
-    @Test
-    void deveriaCriarPersonagemComSucesso() {
-        PersonagemRequest request1 = new PersonagemRequest("Barbaro", TipoClassePersonagem.BARBARO);
-        PersonagemRequest request2 = new PersonagemRequest("Orc", TipoClassePersonagem.ORC);
-        Personagem personagem1 = new BarbaroFactory().criarPersonagem(request1.nome());
-        Personagem personagem2 = new OrcFactory().criarPersonagem(request2.nome());
-
-        Mockito.when(personagemRepository.findByNome(ArgumentMatchers.anyString()))
-            .thenReturn(Optional.empty())
-            .thenReturn(Optional.empty());
-
-        Mockito.when(personagemRepository.save(ArgumentMatchers.any(Personagem.class)))
-            .thenReturn(personagem1)
-            .thenReturn(personagem2);
-
-        PersonagemResponse response1 = personagemService.criarPersonagem(request1);
-        PersonagemResponse response2 = personagemService.criarPersonagem(request2);
-
-        Assertions.assertDoesNotThrow(() -> personagemService.criarPersonagem(request1));
-        Assertions.assertDoesNotThrow(() -> personagemService.criarPersonagem(request2));
-        Assertions.assertEquals(personagem1.getNome(), response1.nome());
-        Assertions.assertEquals(personagem2.getNome(), response2.nome());
-        Assertions.assertNotEquals(response1, response2);
-    }
-
-    @Test
-    void deveriaLancarExcecaoCasoExistaPersonagemComMesmoNome() {
-        PersonagemRequest request = new PersonagemRequest("Barbaro", TipoClassePersonagem.BARBARO);
-        Personagem personagem = new BarbaroFactory().criarPersonagem(request.nome());
-
-        Mockito.when(personagemRepository.findByNome(ArgumentMatchers.anyString()))
-            .thenReturn(Optional.of(personagem));
-
-        Assertions.assertThrows(NomeExistenteException.class, () -> personagemService.criarPersonagem(request));
-    }
-
-    @Test
-    void deveriaLancarExcecaoCasoNaoExistaFactoryParaClasseRequisitada() {
-        PersonagemRequest request = new PersonagemRequest("Barbaro", TipoClassePersonagem.BARBARO);
-
-        Mockito.when(personagemRepository.findByNome(ArgumentMatchers.anyString()))
-            .thenReturn(Optional.empty());
-
-        Mockito.when(factoryConfiguration.getFactory(ArgumentMatchers.any(TipoClassePersonagem.class)))
-            .thenReturn(null);
-
-        Assertions.assertThrows(NotFoundException.class, () -> personagemService.criarPersonagem(request));
-    }
-
-    @Test
-    void deveriaBuscarTodosOsPersonagensSalvosNoBanco() {
-        Personagem personagem1 = new BarbaroFactory().criarPersonagem("Barbaro");
-        Personagem personagem2 = new OrcFactory().criarPersonagem("Orc");
-
-        Mockito.when(personagemRepository.findAll())
-            .thenReturn(List.of(personagem1, personagem2));
-
-        Assertions.assertEquals(2, personagemService.listaTodosPersonagens().size());
-    }
-
-    @Test
-    void deveriaRetornarUsuarioPorNomeComSucesso() {
-        Personagem personagem = new BarbaroFactory().criarPersonagem("Barbaro");
-
-        Mockito.when(personagemRepository.findByNomeOrThrowException(ArgumentMatchers.anyString()))
-            .thenReturn(personagem);
-
-        Assertions.assertEquals("Barbaro", personagemService.buscaPersonagemPorNome("Barbaro").nome());
-    }
-
-    @Test
-    void deveriaLancarExcecaoCasoNaoEncontrePersonagemPorNome() {
-        Mockito.when(personagemRepository.findByNome("Nome Inexistente"))
-            .thenReturn(Optional.empty());
-
-        // Verifica a exceção default lançada pelo método findByNomeOrThrowException do JpaRepositoy
-        Assertions.assertThrows(NullPointerException.class, () -> personagemService.buscaPersonagemPorNome("Nome Inexistente"));
-    }
-
-    @Test
-    void deveriaRetornarUsuarioPorUuidComSucesso() {
-        Personagem personagem = new BarbaroFactory().criarPersonagem("Barbaro");
-
-        Mockito.when(personagemRepository.findByIdWithExceptionIfNotFound(ArgumentMatchers.any()))
-            .thenReturn(personagem);
-
-        Assertions.assertEquals("Barbaro", personagemService.buscaPersonagemPorUuid(personagem.getUuid()).nome());
-    }
-
-    @Test
-    void deveriaLancarExcecaoCasoNaoEncontrePersonagemPorUuid() {
-        UUID randomUuid = UUID.randomUUID();
-
-        Mockito.when(personagemRepository.findById(ArgumentMatchers.any()))
-            .thenReturn(Optional.empty());
-
-        Assertions.assertThrows(NullPointerException.class, () -> personagemService.buscaPersonagemPorUuid(randomUuid));
-    }
-
-    @Test
-    void deveriaAtualizarPersonagemComSucesso() {
-        Personagem personagem = new BarbaroFactory().criarPersonagem("Barbaro");
-        String nomeAnterior = personagem.getNome();
-        PersonagemUpdateRequest request = new PersonagemUpdateRequest("Barbaro Atualizado");
-
-        Mockito.when(personagemRepository.findByIdWithExceptionIfNotFound(ArgumentMatchers.any()))
-            .thenReturn(personagem);
-
-        Mockito.when(personagemRepository.save(ArgumentMatchers.any()))
-            .thenReturn(personagem);
-        PersonagemResponse response = personagemService.atualizaPersonagem(personagem.getUuid(), request);
-
-        Assertions.assertEquals(request.nome(), response.nome());
-        Assertions.assertNotEquals(nomeAnterior, response.nome());
-    }
-
-    @Test
-    void deveriaDeletarPersonagemComSucesso() {
-        Personagem personagem = new BarbaroFactory().criarPersonagem("Barbaro");
-
-        Mockito.when(personagemRepository.findByIdWithExceptionIfNotFound(ArgumentMatchers.any()))
-            .thenReturn(personagem);
-
-        Assertions.assertDoesNotThrow(() -> personagemService.deletaPersonagem(personagem.getUuid()));
-    }
-}
